refactor(024): promote V2 bottom sheet to default, remove V1 and feature flag
BREAKING CHANGE: FireDetailsBottomSheet now uses V2 implementation

- Removed USE_BOTTOM_SHEET_V2 feature flag
- Renamed FireDetailsBottomSheetV2 → FireDetailsBottomSheet
- Deleted original V1 fire_details_bottom_sheet.dart
- Updated MapScreen to use V2 directly (no conditional)
- Tests need updating to match V2 text strings
