+++ lib/widgets/fire_details_bottom_sheet.dart
@@ -1,83 +1,81 @@
 import 'package:flutter/material.dart';
-import 'package:wildfire_mvp_v3/models/fire_incident.dart';
-import 'package:wildfire_mvp_v3/models/location_models.dart';
-import 'package:wildfire_mvp_v3/models/hotspot.dart';
+import 'package:go_router/go_router.dart';
+import 'package:wildfire_mvp_v3/features/map/widgets/map_source_chip.dart';
 import 'package:wildfire_mvp_v3/models/burnt_area.dart';
+import 'package:wildfire_mvp_v3/models/fire_incident.dart';
+import 'package:wildfire_mvp_v3/models/hotspot.dart';
+import 'package:wildfire_mvp_v3/models/location_models.dart';
 import 'package:wildfire_mvp_v3/services/models/fire_risk.dart';
 import 'package:wildfire_mvp_v3/utils/distance_calculator.dart';
-import 'package:wildfire_mvp_v3/features/map/widgets/map_source_chip.dart';
 
-/// Type of fire data being displayed
+/// Display type for fire data in the bottom sheet
 enum FireDataDisplayType {
-  /// Standard fire incident (from FireLocationService)
+  /// Standard fire incident (legacy)
   incident,
 
-  /// Active hotspot from VIIRS satellite
+  /// Live hotspot (VIIRS satellite detection)
   hotspot,
 
-  /// Verified burnt area from MODIS satellite
+  /// Verified burnt area (MODIS confirmed)
   burntArea,
 }
 
-/// Bottom sheet widget displaying comprehensive fire incident details
+/// V2 Bottom sheet for fire details with improved UX
 ///
-/// Shows detection time, source, confidence, FRP, distance, bearing,
-/// and other metadata in a scrollable sheet. Supports loading states,
-/// error handling, and retry functionality.
+/// Key improvements over V1:
+/// - Dynamic header based on data type (not generic "Fire Incident Details")
+/// - Summary card with key info at top
+/// - Plain language descriptions (no jargon)
+/// - Progressive disclosure (land cover hidden by default)
+/// - "Learn More" link to help docs
+/// - Improved time format: "X ago (HH:MM UK time)"
 ///
-/// Example usage:
+/// Usage:
 /// ```dart
-/// // Simple usage with incident data
-/// FireDetailsBottomSheet(
-///   incident: fireIncident,
-///   userLocation: currentLocation,
-///   onClose: () {},
-/// )
-///
-/// // With loading state
-/// FireDetailsBottomSheet(
-///   isLoading: true,
-///   onClose: () {},
-/// )
-///
-/// // With error state
-/// FireDetailsBottomSheet(
-///   errorMessage: 'Failed to load fire details',
-///   onClose: () {},
-///   onRetry: () {},
-/// )
+/// showModalBottomSheet(
+///   context: context,
+///   isScrollControlled: true,
+///   builder: (_) => FireDetailsBottomSheet.fromHotspot(
+///     hotspot: hotspot,
+///     userLocation: userLocation,
+///   ),
+/// );
 /// ```
-///
-/// Constitutional compliance:
-/// - C3 (Accessibility): ≥44dp touch targets, semantic labels, high contrast
-/// - C4 (Transparency): Shows data source, freshness, all detection metadata
 class FireDetailsBottomSheet extends StatelessWidget {
-  /// Fire incident to display (nullable for loading/error states)
+  /// The fire incident being displayed
   final FireIncident? incident;
 
-  /// User's current location for distance/bearing calculation
+  /// User's current GPS location (for distance calculation)
   final LatLng? userLocation;
 
-  /// Optional callback when close button is tapped
+  /// Callback when close button is pressed
   final VoidCallback? onClose;
 
-  /// Optional callback when retry button is tapped (error state only)
+  /// Callback when retry button is pressed (error state only)
   final VoidCallback? onRetry;
 
-  /// Whether to show loading state instead of incident data
-  final bool isLoading;
-
-  /// Error message to display (shows error state when non-null)
-  final String? errorMessage;
-
-  /// Type of fire data being displayed (incident, hotspot, or burntArea)
+  /// What type of fire data we're displaying
   final FireDataDisplayType displayType;
 
-  /// Hotspot source data (when displayType == hotspot)
+  /// Original Hotspot data (when created via fromHotspot factory)
   final Hotspot? hotspot;
 
-  /// BurntArea source data (when displayType == burntArea)
+  /// Original BurntArea data (when created via fromBurntArea factory)
   final BurntArea? burntArea;
+
+  /// Whether the sheet is in loading state
+  final bool isLoading;
+
+  /// Error message to display (null if no error)
+  final String? errorMessage;
+
+  /// Optional callback for "Learn More" action
+  /// If null, defaults to navigating to appropriate help doc
+  final VoidCallback? onLearnMore;
+
+  /// Whether the user's location was set manually (vs GPS)
+  /// Used to display "Your Location (Manual)" vs "Your Location (GPS)"
+  final bool isManualLocation;
 
   const FireDetailsBottomSheet({
     super.key,
@@ -85,17 +83,16 @@ class FireDetailsBottomSheet extends StatelessWidget {
     this.userLocation,
     this.onClose,
     this.onRetry,
-    this.isLoading = false,
-    this.errorMessage,
     this.displayType = FireDataDisplayType.incident,
     this.hotspot,
     this.burntArea,
+    this.isLoading = false,
+    this.errorMessage,
+    this.onLearnMore,
+    this.isManualLocation = false,
   });
 
   /// Factory constructor for displaying a Hotspot
-  ///
-  /// Converts Hotspot to FireIncident for display and sets displayType.
-  /// The [freshness] parameter indicates whether data is live or mock.
   factory FireDetailsBottomSheet.fromHotspot({
     Key? key,
     required Hotspot hotspot,
@@ -103,6 +100,8 @@ class FireDetailsBottomSheet extends StatelessWidget {
     VoidCallback? onClose,
     VoidCallback? onRetry,
     Freshness freshness = Freshness.live,
+    VoidCallback? onLearnMore,
+    bool isManualLocation = false,
   }) {
     // Convert Hotspot to FireIncident for rendering
     final incident = FireIncident(
@@ -126,13 +125,12 @@ class FireDetailsBottomSheet extends StatelessWidget {
       onRetry: onRetry,
       displayType: FireDataDisplayType.hotspot,
       hotspot: hotspot,
+      onLearnMore: onLearnMore,
+      isManualLocation: isManualLocation,
     );
   }
 
   /// Factory constructor for displaying a BurntArea
-  ///
-  /// Converts BurntArea to FireIncident for display and sets displayType.
-  /// The [freshness] parameter indicates whether data is live or mock.
   factory FireDetailsBottomSheet.fromBurntArea({
     Key? key,
     required BurntArea burntArea,
@@ -140,6 +138,8 @@ class FireDetailsBottomSheet extends StatelessWidget {
     VoidCallback? onClose,
     VoidCallback? onRetry,
     Freshness freshness = Freshness.live,
+    VoidCallback? onLearnMore,
+    bool isManualLocation = false,
   }) {
     // Convert BurntArea to FireIncident for rendering
     final incident = FireIncident(
@@ -163,6 +163,8 @@ class FireDetailsBottomSheet extends StatelessWidget {
       onRetry: onRetry,
       displayType: FireDataDisplayType.burntArea,
       burntArea: burntArea,
+      onLearnMore: onLearnMore,
+      isManualLocation: isManualLocation,
     );
   }
 
@@ -172,15 +174,14 @@ class FireDetailsBottomSheet extends StatelessWidget {
     final colorScheme = theme.colorScheme;
 
     return DraggableScrollableSheet(
-      initialChildSize: 0.4,
+      initialChildSize: 0.45,
       minChildSize: 0.2,
-      maxChildSize: 0.8,
+      maxChildSize: 0.85,
       expand: false,
       builder: (context, scrollController) {
         return Container(
           decoration: BoxDecoration(
             color: colorScheme.surfaceContainerHighest,
-            //color: colorScheme.surface,
             borderRadius: const BorderRadius.only(
               topLeft: Radius.circular(24),
               topRight: Radius.circular(24),
@@ -198,10 +199,8 @@ class FireDetailsBottomSheet extends StatelessWidget {
               // Drag handle
               _buildDragHandle(context),
 
-              // Header with title and close button
+              // Dynamic header based on type
               _buildHeader(context),
-
-              const Divider(height: 1),
 
               // Scrollable content based on state
               Expanded(
@@ -244,12 +243,12 @@ class FireDetailsBottomSheet extends StatelessWidget {
       children: [
         const SizedBox(height: 32),
         Semantics(
-          label: 'Loading fire incident details',
+          label: 'Loading fire details',
           child: const CircularProgressIndicator(),
         ),
         const SizedBox(height: 16),
         Text(
-          'Loading fire details...',
+          'Loading details...',
           style: Theme.of(context).textTheme.bodyMedium,
           textAlign: TextAlign.center,
         ),
@@ -270,7 +269,7 @@ class FireDetailsBottomSheet extends StatelessWidget {
         ),
         const SizedBox(height: 16),
         Text(
-          'Failed to Load Fire Details',
+          'Failed to Load Details',
           style: Theme.of(context).textTheme.titleMedium?.copyWith(
                 color: Theme.of(context).colorScheme.error,
                 fontWeight: FontWeight.bold,
@@ -298,237 +297,390 @@ class FireDetailsBottomSheet extends StatelessWidget {
   }
 
   Widget _buildLoadedState(BuildContext context) {
-    final inc = incident!; // Safe to unwrap, checked in _buildContent
+    final inc = incident!;
 
+    // Route to the appropriate layout based on display type
+    return switch (displayType) {
+      FireDataDisplayType.burntArea => _buildBurntAreaLayout(context, inc),
+      FireDataDisplayType.hotspot => _buildHotspotLayout(context, inc),
+      FireDataDisplayType.incident => _buildIncidentLayout(context, inc),
+    };
+  }
+
+  /// Build layout for Burnt Area bottom sheet
+  Widget _buildBurntAreaLayout(BuildContext context, FireIncident inc) {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        // Educational label for hotspot/burntArea types
-        if (displayType != FireDataDisplayType.incident)
-          _buildEducationalLabel(context),
+        // Supporting text with inline learn more link
+        _buildSupportingTextWithLearnMore(context),
 
-        // Data source and demo chips
-        _buildChips(inc),
+        const SizedBox(height: 16),
 
-        const SizedBox(height: 20),
-
-        // Simplification notice for burnt areas
-        if (displayType == FireDataDisplayType.burntArea &&
-            burntArea != null &&
-            burntArea!.isSimplified)
-          _buildSimplificationNotice(context),
-
-        // Land cover breakdown for burnt areas
-        if (displayType == FireDataDisplayType.burntArea &&
-            burntArea?.landCoverBreakdown != null &&
-            burntArea!.landCoverBreakdown!.isNotEmpty)
-          _buildLandCoverSection(context),
-
-        // Location section
-        _InfoSection(
-          title: 'Location',
-          icon: Icons.place_outlined,
-          children: [
-            // User's GPS location status
-            _buildDetailRow(
-              context: context,
-              icon: userLocation != null ? Icons.gps_fixed : Icons.gps_off,
-              label: 'Your location',
-              value: userLocation != null
-                  ? '${userLocation!.latitude.toStringAsFixed(2)}, ${userLocation!.longitude.toStringAsFixed(2)} (GPS)'
-                  : 'Unknown (GPS unavailable)',
-              semanticLabel: userLocation != null
-                  ? 'Your GPS location: ${userLocation!.latitude.toStringAsFixed(2)} latitude, ${userLocation!.longitude.toStringAsFixed(2)} longitude'
-                  : 'Your GPS location is unavailable',
+        // Key Metrics section
+        _buildKeyMetricsSection(
+          context: context,
+          rows: [
+            if (inc.areaHectares != null)
+              _KeyMetricRow(
+                icon: Icons.square_foot,
+                label: 'Est. burned area',
+                value: '${inc.areaHectares!.toStringAsFixed(1)} ha',
+              ),
+            _KeyMetricRow(
+              icon: Icons.access_time,
+              label: 'When detected',
+              value:
+                  _formatRelativeTimeWithClock(inc.detectedAt ?? inc.timestamp),
             ),
-            // Distance and direction (only if GPS available)
-            _buildDetailRow(
-              context: context,
-              icon: Icons.social_distance,
-              label: 'Distance & direction',
-              value: userLocation != null
-                  ? '${DistanceCalculator.formatDistanceAndDirection(userLocation!, inc.location)} from your location'
-                  : 'Unable to calculate (GPS required)',
-              semanticLabel: userLocation != null
-                  ? 'Fire is ${DistanceCalculator.formatDistanceAndDirection(userLocation!, inc.location)} from your GPS location'
-                  : 'Distance unavailable because GPS location is unknown',
-            ),
-            _buildDetailRow(
-              context: context,
+            _KeyMetricRow(
               icon: Icons.location_on,
               label: 'Fire coordinates',
               value:
                   '${inc.location.latitude.toStringAsFixed(4)}, ${inc.location.longitude.toStringAsFixed(4)}',
-              semanticLabel:
-                  'Fire coordinates: ${inc.location.latitude.toStringAsFixed(4)} latitude, ${inc.location.longitude.toStringAsFixed(4)} longitude',
             ),
-          ],
-        ),
-
-        const SizedBox(height: 16),
-
-        // Fire characteristics section
-        _InfoSection(
-          title: 'Fire characteristics',
-          icon: Icons.local_fire_department_outlined,
-          children: [
-            // Fire ID first
-            _buildDetailRow(
-              context: context,
-              icon: Icons.fingerprint,
-              label: 'Fire ID',
-              value: inc.id,
-              semanticLabel: 'Fire ID: ${inc.id}',
-            ),
-            if (inc.areaHectares != null)
-              _buildDetailRow(
-                context: context,
-                icon: Icons.square_foot,
-                label: 'Estimated burned area',
-                value: '${inc.areaHectares!.toStringAsFixed(1)} hectares (ha)',
-                semanticLabel:
-                    'Estimated area: ${inc.areaHectares!.toStringAsFixed(1)} hectares',
-              ),
-            if (inc.frp != null)
-              _buildDetailRow(
-                context: context,
-                icon: Icons.power,
-                label: 'Fire power (FRP)',
-                value: '${inc.frp!.toStringAsFixed(0)} MW',
-                semanticLabel:
-                    'Fire radiative power: ${inc.frp!.toStringAsFixed(0)} megawatts',
-              ),
-            // Only show intensity for hotspots where it's derived from FRP
-            // Burnt areas don't have an intensity/risk field in the API
-            if (displayType == FireDataDisplayType.hotspot)
-              _buildDetailRow(
-                context: context,
-                icon: Icons.whatshot,
-                label: 'Fire intensity',
-                value: _formatIntensity(inc.intensity),
-                semanticLabel:
-                    'Fire intensity: ${_formatIntensity(inc.intensity)}',
-              ),
-            if (inc.confidence != null)
-              _buildDetailRow(
-                context: context,
-                icon: Icons.verified,
-                label: 'Detection confidence',
-                value: '${inc.confidence!.toStringAsFixed(0)}%',
-                semanticLabel:
-                    'Detection confidence: ${inc.confidence!.toStringAsFixed(0)} percent',
+            if (userLocation != null)
+              _KeyMetricRow(
+                icon: Icons.social_distance,
+                label: 'Distance from your location',
+                value:
+                    _formatDistanceWithDirection(userLocation!, inc.location),
               ),
           ],
         ),
 
         const SizedBox(height: 16),
 
-        // Detection and Source section
-        _InfoSection(
-          title: 'Detection & source',
-          icon: Icons.access_time,
+        // More Details expandable section
+        _ExpandableDetailsSection(
+          title: 'More details',
           children: [
-            _buildDetailRow(
-              context: context,
-              icon: Icons.access_time,
-              label: 'Detected',
-              // Use relative time for hotspots, absolute for others
-              value: displayType == FireDataDisplayType.hotspot
-                  ? _formatRelativeTime(inc.detectedAt ?? inc.timestamp)
-                  : _formatDateTime(inc.detectedAt ?? inc.timestamp),
-              semanticLabel: displayType == FireDataDisplayType.hotspot
-                  ? _formatRelativeTime(inc.detectedAt ?? inc.timestamp)
-                  : 'Detected at ${_formatDateTime(inc.detectedAt ?? inc.timestamp)}',
-            ),
-            // Show fire season for burnt areas (helps verify correct season filter)
-            if (displayType == FireDataDisplayType.burntArea &&
-                burntArea != null)
+            if (userLocation != null)
               _buildDetailRow(
                 context: context,
-                icon: Icons.calendar_month,
-                label: 'Fire season',
-                value: '${burntArea!.seasonYear}',
-                semanticLabel: 'Fire season: ${burntArea!.seasonYear}',
+                icon: isManualLocation ? Icons.edit_location : Icons.gps_fixed,
+                label: isManualLocation
+                    ? 'Your Location (Manual)'
+                    : 'Your Location (GPS)',
+                value:
+                    '${userLocation!.latitude.toStringAsFixed(2)}, ${userLocation!.longitude.toStringAsFixed(2)}',
               ),
             _buildDetailRow(
               context: context,
               icon: Icons.source,
               label: 'Data source',
-              // Show accurate data source based on display type and freshness
               value: _formatDataSourceForType(inc.source, inc.freshness),
-              semanticLabel:
-                  'Data source: ${_formatDataSourceForType(inc.source, inc.freshness)}',
             ),
             _buildDetailRow(
               context: context,
               icon: Icons.satellite_alt,
               label: 'Sensor',
-              value: inc.sensorSource ?? 'Unknown sensor',
-              semanticLabel:
-                  'Sensor source: ${inc.sensorSource ?? 'Unknown sensor'}',
+              value: _formatSensorName(inc.sensorSource),
             ),
-            if (inc.lastUpdate != null)
-              _buildDetailRow(
-                context: context,
-                icon: Icons.update,
-                label: 'Last updated',
-                value: _formatDateTime(inc.lastUpdate!),
-                semanticLabel:
-                    'Last updated at ${_formatDateTime(inc.lastUpdate!)}',
+            _buildDetailRow(
+              context: context,
+              icon: Icons.fingerprint,
+              label: 'Fire ID',
+              value: inc.id,
+            ),
+            // Land cover breakdown
+            if (burntArea?.landCoverBreakdown != null &&
+                burntArea!.landCoverBreakdown!.isNotEmpty)
+              _buildLandCoverSection(context),
+            // Polygon simplification notice
+            if (burntArea != null && burntArea!.isSimplified)
+              _buildSimplificationNotice(context),
+          ],
+        ),
+
+        const SizedBox(height: 16),
+
+        // Safety note at bottom
+        _buildSafetyText(context),
+
+        const SizedBox(height: 16),
+      ],
+    );
+  }
+
+  /// Build layout for Hotspot bottom sheet
+  Widget _buildHotspotLayout(BuildContext context, FireIncident inc) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        // Supporting text with inline learn more link
+        _buildSupportingTextWithLearnMore(context),
+
+        const SizedBox(height: 16),
+
+        // Key Metrics section
+        _buildKeyMetricsSection(
+          context: context,
+          rows: [
+            _KeyMetricRow(
+              icon: Icons.location_on,
+              label: 'Fire coordinates',
+              value:
+                  '${inc.location.latitude.toStringAsFixed(4)}, ${inc.location.longitude.toStringAsFixed(4)}',
+            ),
+            _KeyMetricRow(
+              icon: Icons.access_time,
+              label: 'When detected',
+              value:
+                  _formatRelativeTimeWithClock(inc.detectedAt ?? inc.timestamp),
+            ),
+            if (userLocation != null)
+              _KeyMetricRow(
+                icon: Icons.social_distance,
+                label: 'Distance from your location',
+                value:
+                    _formatDistanceWithDirection(userLocation!, inc.location),
+              ),
+            if (inc.confidence != null)
+              _KeyMetricRow(
+                icon: Icons.verified_outlined,
+                label: 'Satellite confidence',
+                value: '${inc.confidence!.toStringAsFixed(0)}%',
+              ),
+            if (inc.intensity.isNotEmpty)
+              _KeyMetricRow(
+                icon: Icons.whatshot,
+                label: 'Fire intensity',
+                value: _formatIntensity(inc.intensity),
               ),
           ],
         ),
 
         const SizedBox(height: 16),
 
-        // Safety warning text
-        Text(
-          'These details come from satellite detections and may lag behind real-world conditions. '
-          'If you are in immediate danger, call 999 without delay.',
-          style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                color: Theme.of(context).colorScheme.onSurfaceVariant,
+        // More Details expandable section
+        _ExpandableDetailsSection(
+          title: 'More details',
+          children: [
+            if (userLocation != null)
+              _buildDetailRow(
+                context: context,
+                icon: isManualLocation ? Icons.edit_location : Icons.gps_fixed,
+                label: isManualLocation
+                    ? 'Your Location (Manual)'
+                    : 'Your Location (GPS)',
+                value:
+                    '${userLocation!.latitude.toStringAsFixed(2)}, ${userLocation!.longitude.toStringAsFixed(2)}',
               ),
+            _buildDetailRow(
+              context: context,
+              icon: Icons.source,
+              label: 'Data source',
+              value: _formatDataSourceForType(inc.source, inc.freshness),
+            ),
+            _buildDetailRow(
+              context: context,
+              icon: Icons.satellite_alt,
+              label: 'Sensor',
+              value: _formatSensorName(inc.sensorSource),
+            ),
+            _buildDetailRow(
+              context: context,
+              icon: Icons.fingerprint,
+              label: 'Ref ID',
+              value: inc.id,
+            ),
+            if (inc.frp != null)
+              _buildDetailRow(
+                context: context,
+                icon: Icons.bolt,
+                label: 'Fire Radiative Power',
+                value: '${inc.frp!.toStringAsFixed(1)} MW',
+              ),
+          ],
         ),
+
+        const SizedBox(height: 16),
+
+        // Safety note at bottom
+        _buildSafetyText(context),
+
         const SizedBox(height: 16),
       ],
     );
   }
 
-  Widget _buildDragHandle(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.symmetric(vertical: 8),
-      width: 40,
-      height: 4,
-      decoration: BoxDecoration(
-        color: Theme.of(context).colorScheme.outlineVariant,
-        borderRadius: BorderRadius.circular(999),
-      ),
+  /// Build layout for legacy Incident type (simplified, future use)
+  Widget _buildIncidentLayout(BuildContext context, FireIncident inc) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        // Key Metrics section
+        _buildKeyMetricsSection(
+          context: context,
+          rows: [
+            _KeyMetricRow(
+              icon: Icons.location_on,
+              label: 'Fire coordinates',
+              value:
+                  '${inc.location.latitude.toStringAsFixed(4)}, ${inc.location.longitude.toStringAsFixed(4)}',
+            ),
+            _KeyMetricRow(
+              icon: Icons.access_time,
+              label: 'When detected',
+              value:
+                  _formatRelativeTimeWithClock(inc.detectedAt ?? inc.timestamp),
+            ),
+            if (userLocation != null)
+              _KeyMetricRow(
+                icon: Icons.social_distance,
+                label: 'Distance from your location',
+                value:
+                    _formatDistanceWithDirection(userLocation!, inc.location),
+              ),
+          ],
+        ),
+
+        const SizedBox(height: 16),
+
+        // More Details expandable section
+        _ExpandableDetailsSection(
+          title: 'More details',
+          children: [
+            if (userLocation != null)
+              _buildDetailRow(
+                context: context,
+                icon: isManualLocation ? Icons.edit_location : Icons.gps_fixed,
+                label: isManualLocation
+                    ? 'Your Location (Manual)'
+                    : 'Your Location (GPS)',
+                value:
+                    '${userLocation!.latitude.toStringAsFixed(2)}, ${userLocation!.longitude.toStringAsFixed(2)}',
+              ),
+            _buildDetailRow(
+              context: context,
+              icon: Icons.source,
+              label: 'Data source',
+              value: _formatDataSourceForType(inc.source, inc.freshness),
+            ),
+            if (inc.id.isNotEmpty)
+              _buildDetailRow(
+                context: context,
+                icon: Icons.fingerprint,
+                label: 'Ref ID',
+                value: inc.id,
+              ),
+          ],
+        ),
+
+        const SizedBox(height: 16),
+
+        // Safety note at bottom
+        _buildSafetyText(context),
+
+        const SizedBox(height: 16),
+      ],
     );
   }
 
+  /// Build Key Metrics section with title and card
+  Widget _buildKeyMetricsSection({
+    required BuildContext context,
+    required List<_KeyMetricRow> rows,
+  }) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+    final isDark = theme.brightness == Brightness.dark;
+
+    // Filter out null rows
+    final validRows = rows.where((r) => true).toList();
+    if (validRows.isEmpty) return const SizedBox.shrink();
+
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        // Section title
+        Row(
+          children: [
+            Icon(Icons.bar_chart,
+                size: 18, color: colorScheme.onSurfaceVariant),
+            const SizedBox(width: 6),
+            Text(
+              'Key metrics',
+              style: theme.textTheme.titleSmall?.copyWith(
+                fontWeight: FontWeight.w600,
+                color: colorScheme.onSurfaceVariant,
+              ),
+            ),
+          ],
+        ),
+        const SizedBox(height: 8),
+        // Card container
+        Container(
+          decoration: BoxDecoration(
+            color: isDark
+                ? colorScheme.surfaceContainerHigh
+                : colorScheme.surfaceContainerLowest,
+            borderRadius: BorderRadius.circular(12),
+            border: Border.all(color: colorScheme.outlineVariant),
+          ),
+          padding: const EdgeInsets.all(12),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              for (int i = 0; i < validRows.length; i++) ...[
+                if (i > 0) const SizedBox(height: 12),
+                _buildDetailRow(
+                  context: context,
+                  icon: validRows[i].icon,
+                  label: validRows[i].label,
+                  value: validRows[i].value,
+                  semanticLabel: '${validRows[i].label}: ${validRows[i].value}',
+                ),
+              ],
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+
+  /// Build dynamic header based on display type
   Widget _buildHeader(BuildContext context) {
     final theme = Theme.of(context);
     final colorScheme = theme.colorScheme;
     final textColor =
         theme.textTheme.titleLarge?.color ?? colorScheme.onSurface;
 
+    // Dynamic title and icon based on type
+    // C2: Use less assertive title for hotspots ("Satellite Hotspot" not "Active Hotspot")
+    final (icon, title) = switch (displayType) {
+      FireDataDisplayType.hotspot => (
+          Icons.local_fire_department,
+          'Satellite Hotspot',
+        ),
+      FireDataDisplayType.burntArea => (
+          Icons.layers,
+          'Burnt Area',
+        ),
+      FireDataDisplayType.incident => (
+          Icons.local_fire_department,
+          'Fire Details',
+        ),
+    };
+
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 16),
       child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
+          // Title section (left)
           Row(
             mainAxisSize: MainAxisSize.min,
             children: [
               Icon(
-                Icons.local_fire_department,
+                icon,
                 size: 24,
                 color: textColor,
                 semanticLabel: '',
               ),
               const SizedBox(width: 12),
               Text(
-                'Fire Incident Details',
+                title,
                 style: theme.textTheme.titleLarge?.copyWith(
                   fontWeight: FontWeight.bold,
                   color: textColor,
@@ -536,9 +688,19 @@ class FireDetailsBottomSheet extends StatelessWidget {
               ),
             ],
           ),
+          // Demo chip centered in remaining space
+          if (incident != null)
+            Expanded(
+              child: Center(
+                child: _buildHeaderChip(context),
+              ),
+            )
+          else
+            const Spacer(),
+          // Close button (right)
           Semantics(
             button: true,
-            label: 'Close fire details',
+            label: 'Close details',
             child: IconButton(
               icon: Icon(Icons.close, color: colorScheme.onSurfaceVariant),
               onPressed: () {
@@ -557,8 +719,93 @@ class FireDetailsBottomSheet extends StatelessWidget {
     );
   }
 
-  Widget _buildChips(FireIncident inc) {
-    return MapSourceChip(source: inc.freshness, lastUpdated: inc.timestamp);
+  Widget _buildDragHandle(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.symmetric(vertical: 8),
+      width: 40,
+      height: 4,
+      decoration: BoxDecoration(
+        color: Theme.of(context).colorScheme.outlineVariant,
+        borderRadius: BorderRadius.circular(999),
+      ),
+    );
+  }
+
+  /// Build compact chip for header (just source label)
+  Widget _buildHeaderChip(BuildContext context) {
+    if (incident == null) return const SizedBox.shrink();
+    return MapSourceChip(
+      source: incident!.freshness,
+      lastUpdated: incident!.timestamp,
+    );
+  }
+
+  /// Build supporting text with inline "Learn more" link at the end
+  Widget _buildSupportingTextWithLearnMore(BuildContext context) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+
+    final (description, helpDocId) = switch (displayType) {
+      FireDataDisplayType.hotspot => (
+          'A satellite detected unusual heat here. It could be a wildfire, controlled burning, or another heat source.',
+          'hotspots',
+        ),
+      FireDataDisplayType.burntArea => (
+          'This outline indicates land that appears to have burned earlier in the season.',
+          'burnt-area',
+        ),
+      FireDataDisplayType.incident => (null, 'hotspots'),
+    };
+
+    if (description == null) return const SizedBox.shrink();
+
+    return Semantics(
+      label: '$description Learn more.',
+      child: Text.rich(
+        TextSpan(
+          style: theme.textTheme.bodyMedium?.copyWith(
+            color: colorScheme.onSurfaceVariant,
+            height: 1.4,
+          ),
+          children: [
+            TextSpan(text: '$description '),
+            WidgetSpan(
+              alignment: PlaceholderAlignment.baseline,
+              baseline: TextBaseline.alphabetic,
+              child: GestureDetector(
+                onTap: () {
+                  if (onLearnMore != null) {
+                    onLearnMore!();
+                  } else {
+                    context.push('/help/doc/$helpDocId');
+                  }
+                },
+                child: Row(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    Icon(
+                      Icons.help_outline,
+                      size: 16,
+                      color: colorScheme.primary,
+                    ),
+                    const SizedBox(width: 4),
+                    Text(
+                      'Learn more',
+                      style: theme.textTheme.bodyMedium?.copyWith(
+                        color: colorScheme.primary,
+                        fontWeight: FontWeight.w500,
+                        decoration: TextDecoration.underline,
+                        decorationColor: colorScheme.primary,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
   }
 
   Widget _buildDetailRow({
@@ -598,7 +845,6 @@ class FireDetailsBottomSheet extends StatelessWidget {
                   child: Text(
                     value,
                     style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                          //fontWeight: FontWeight.w500,
                           fontWeight: FontWeight.w600,
                           fontSize: 15,
                         ),
@@ -612,199 +858,7 @@ class FireDetailsBottomSheet extends StatelessWidget {
     );
   }
 
-  String _formatDateTime(DateTime dateTime) {
-    final utcTime = dateTime.toUtc();
-    final now = DateTime.now().toUtc();
-
-    // Format time as 12-hour with AM/PM
-    final hour12 = utcTime.hour > 12
-        ? utcTime.hour - 12
-        : (utcTime.hour == 0 ? 12 : utcTime.hour);
-    final amPm = utcTime.hour >= 12 ? 'PM' : 'AM';
-    final timeString =
-        '$hour12:${utcTime.minute.toString().padLeft(2, '0')} $amPm';
-
-    // Check if today
-    if (now.year == utcTime.year &&
-        now.month == utcTime.month &&
-        now.day == utcTime.day) {
-      return 'Today at $timeString UTC';
-    }
-
-    // Check if yesterday
-    final yesterday = now.subtract(const Duration(days: 1));
-    if (yesterday.year == utcTime.year &&
-        yesterday.month == utcTime.month &&
-        yesterday.day == utcTime.day) {
-      return 'Yesterday at $timeString UTC';
-    }
-
-    // Otherwise show date with year
-    const months = [
-      'Jan',
-      'Feb',
-      'Mar',
-      'Apr',
-      'May',
-      'Jun',
-      'Jul',
-      'Aug',
-      'Sep',
-      'Oct',
-      'Nov',
-      'Dec',
-    ];
-    final monthName = months[utcTime.month - 1];
-    return '${utcTime.day} $monthName ${utcTime.year} at $timeString UTC';
-  }
-
-  String _formatIntensity(String intensity) {
-    return intensity.substring(0, 1).toUpperCase() + intensity.substring(1);
-  }
-
-  /// Format data source with accurate service name based on display type
-  ///
-  /// Hotspots come from GWIS WMS (VIIRS layer)
-  /// Burnt areas come from EFFIS WFS (MODIS layer)
-  String _formatDataSourceForType(DataSource source, Freshness freshness) {
-    // For hotspots, the actual source is GWIS (not EFFIS)
-    // Check freshness for mock detection since source is always effis from factory
-    if (displayType == FireDataDisplayType.hotspot) {
-      if (source == DataSource.mock || freshness == Freshness.mock) {
-        return 'Demo Data';
-      }
-      return 'GWIS (EC JRC)'; // Global Wildfire Information System
-    }
-
-    // For burnt areas, it's genuinely EFFIS
-    // Check freshness for mock detection since source is always effis from factory
-    if (displayType == FireDataDisplayType.burntArea) {
-      if (source == DataSource.mock || freshness == Freshness.mock) {
-        return 'Demo Data';
-      }
-      return 'EFFIS (EC JRC)'; // European Forest Fire Information System
-    }
-
-    // Fallback for legacy incidents
-    switch (source) {
-      case DataSource.effis:
-        return 'EFFIS';
-      case DataSource.sepa:
-        return 'SEPA';
-      case DataSource.cache:
-        return 'Cached';
-      case DataSource.mock:
-        return 'Demo Data';
-    }
-  }
-
-  /// Build educational label for hotspot or burnt area types
-  Widget _buildEducationalLabel(BuildContext context) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-
-    final (icon, title, description) = switch (displayType) {
-      FireDataDisplayType.hotspot => (
-          Icons.local_fire_department,
-          'Active Hotspot',
-          'Satellite-detected thermal anomaly indicating possible active fire.',
-        ),
-      FireDataDisplayType.burntArea => (
-          Icons.layers,
-          'Verified Burnt Area',
-          'MODIS satellite-confirmed area affected by fire this season.',
-        ),
-      FireDataDisplayType.incident => (Icons.info_outline, '', ''),
-    };
-
-    if (title.isEmpty) return const SizedBox.shrink();
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 16),
-      child: Semantics(
-        label: '$title. $description',
-        child: Container(
-          padding: const EdgeInsets.all(12),
-          decoration: BoxDecoration(
-            color: colorScheme.primaryContainer.withValues(alpha: 0.3),
-            borderRadius: BorderRadius.circular(12),
-            border: Border.all(
-              color: colorScheme.primary.withValues(alpha: 0.3),
-            ),
-          ),
-          child: Row(
-            children: [
-              Icon(icon, color: colorScheme.primary, size: 24),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      title,
-                      style: theme.textTheme.titleSmall?.copyWith(
-                        fontWeight: FontWeight.bold,
-                        color: colorScheme.primary,
-                      ),
-                    ),
-                    const SizedBox(height: 2),
-                    Text(
-                      description,
-                      style: theme.textTheme.bodySmall?.copyWith(
-                        color: colorScheme.onSurfaceVariant,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  /// Build simplification notice for burnt areas
-  Widget _buildSimplificationNotice(BuildContext context) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final originalCount = burntArea?.originalPointCount ?? 0;
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 16),
-      child: Semantics(
-        label: 'Polygon simplified from $originalCount points for display',
-        child: Container(
-          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-          decoration: BoxDecoration(
-            color: colorScheme.surfaceContainerHighest,
-            borderRadius: BorderRadius.circular(8),
-          ),
-          child: Row(
-            children: [
-              Icon(
-                Icons.compress,
-                size: 16,
-                color: colorScheme.onSurfaceVariant,
-              ),
-              const SizedBox(width: 8),
-              Expanded(
-                child: Text(
-                  'Simplified boundary (from $originalCount points)',
-                  style: theme.textTheme.bodySmall?.copyWith(
-                    color: colorScheme.onSurfaceVariant,
-                    fontStyle: FontStyle.italic,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  /// Build land cover breakdown section for burnt areas
+  /// Build land cover breakdown section
   Widget _buildLandCoverSection(BuildContext context) {
     final theme = Theme.of(context);
     final colorScheme = theme.colorScheme;
@@ -815,11 +869,25 @@ class FireDetailsBottomSheet extends StatelessWidget {
       ..sort((a, b) => b.value.compareTo(a.value));
 
     return Padding(
-      padding: const EdgeInsets.only(bottom: 16),
-      child: _InfoSection(
-        title: 'Land Cover Affected',
-        icon: Icons.terrain,
+      padding: const EdgeInsets.only(bottom: 12),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
         children: [
+          Row(
+            children: [
+              Icon(Icons.terrain,
+                  size: 18, color: colorScheme.onSurfaceVariant),
+              const SizedBox(width: 8),
+              Text(
+                'Land types affected',
+                style: theme.textTheme.bodySmall?.copyWith(
+                  color: colorScheme.onSurfaceVariant,
+                  fontWeight: FontWeight.w500,
+                ),
+              ),
+            ],
+          ),
+          const SizedBox(height: 8),
           for (final entry in sortedEntries)
             _buildLandCoverBar(
               context: context,
@@ -879,12 +947,161 @@ class FireDetailsBottomSheet extends StatelessWidget {
     );
   }
 
+  Widget _buildSimplificationNotice(BuildContext context) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+    final originalCount = burntArea?.originalPointCount ?? 0;
+
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 12),
+      child: Semantics(
+        label: 'Boundary simplified from $originalCount points for display',
+        child: Container(
+          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+          decoration: BoxDecoration(
+            color: colorScheme.surfaceContainerHighest,
+            borderRadius: BorderRadius.circular(8),
+          ),
+          child: Row(
+            children: [
+              Icon(
+                Icons.compress,
+                size: 16,
+                color: colorScheme.onSurfaceVariant,
+              ),
+              const SizedBox(width: 8),
+              Expanded(
+                child: Text(
+                  'Boundary simplified (from $originalCount points)',
+                  style: theme.textTheme.bodySmall?.copyWith(
+                    color: colorScheme.onSurfaceVariant,
+                    fontStyle: FontStyle.italic,
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  /// Build safety warning text
+  Widget _buildSafetyText(BuildContext context) {
+    return Text(
+      'Satellite data may lag behind real conditions. '
+      'If you are in immediate danger, call 999.',
+      style: Theme.of(context).textTheme.bodySmall?.copyWith(
+            color: Theme.of(context).colorScheme.onSurfaceVariant,
+          ),
+    );
+  }
+
+  // ─────────────────────────────────────────────────────────────
+  // Formatting helpers
+  // ─────────────────────────────────────────────────────────────
+
+  /// Format distance with direction for summary card (C1: direction is useful)
+  String _formatDistanceWithDirection(LatLng from, LatLng to) {
+    // Use the existing calculator which returns "X.X km E" format
+    return DistanceCalculator.formatDistanceAndDirection(from, to);
+  }
+
+  /// Format time as "X ago (HH:MM UK time)"
+  String _formatRelativeTimeWithClock(DateTime detectedAt) {
+    final now = DateTime.now().toUtc();
+    final difference = now.difference(detectedAt.toUtc());
+
+    // Format relative part
+    String relativePart;
+    if (difference.inMinutes < 60) {
+      final minutes = difference.inMinutes;
+      relativePart = '$minutes ${minutes == 1 ? 'minute' : 'minutes'} ago';
+    } else if (difference.inHours < 24) {
+      final hours = difference.inHours;
+      relativePart = '$hours ${hours == 1 ? 'hour' : 'hours'} ago';
+    } else if (difference.inDays < 7) {
+      final days = difference.inDays;
+      relativePart = '$days ${days == 1 ? 'day' : 'days'} ago';
+    } else {
+      // For very old data, just show date
+      return _formatDateOnly(detectedAt);
+    }
+
+    // Format UK time (local time for UK users)
+    final ukTime = detectedAt.toLocal();
+    final hour = ukTime.hour.toString().padLeft(2, '0');
+    final minute = ukTime.minute.toString().padLeft(2, '0');
+    final clockPart = '$hour:$minute UK time';
+
+    return '$relativePart ($clockPart)';
+  }
+
+  String _formatDateOnly(DateTime dateTime) {
+    const months = [
+      'Jan',
+      'Feb',
+      'Mar',
+      'Apr',
+      'May',
+      'Jun',
+      'Jul',
+      'Aug',
+      'Sep',
+      'Oct',
+      'Nov',
+      'Dec',
+    ];
+    final d = dateTime.toLocal();
+    return '${d.day} ${months[d.month - 1]} ${d.year}';
+  }
+
+  String _formatIntensity(String intensity) {
+    return intensity.substring(0, 1).toUpperCase() + intensity.substring(1);
+  }
+
+  String _formatSensorName(String? sensor) {
+    if (sensor == null) return 'Unknown';
+    return switch (sensor.toUpperCase()) {
+      'VIIRS' => 'VIIRS satellite',
+      'MODIS' => 'MODIS satellite',
+      _ => sensor,
+    };
+  }
+
+  String _formatDataSourceForType(DataSource source, Freshness freshness) {
+    if (displayType == FireDataDisplayType.hotspot) {
+      if (source == DataSource.mock || freshness == Freshness.mock) {
+        return 'Demo Data';
+      }
+      return 'GWIS (EC JRC)';
+    }
+
+    if (displayType == FireDataDisplayType.burntArea) {
+      if (source == DataSource.mock || freshness == Freshness.mock) {
+        return 'Demo Data';
+      }
+      return 'EFFIS (EC JRC)';
+    }
+
+    switch (source) {
+      case DataSource.effis:
+        return 'EFFIS';
+      case DataSource.sepa:
+        return 'SEPA';
+      case DataSource.cache:
+        return 'Cached';
+      case DataSource.mock:
+        return 'Demo Data';
+    }
+  }
+
   String _formatLandCoverLabel(String key) {
     return switch (key.toLowerCase()) {
       'forest' => 'Forest',
       'shrubland' => 'Shrubland',
       'grassland' => 'Grassland',
-      'agriculture' => 'Agriculture',
+      'agriculture' => 'Farmland',
       'wetland' => 'Wetland',
       'urban' => 'Urban',
       'other' => 'Other',
@@ -903,76 +1120,174 @@ class FireDetailsBottomSheet extends StatelessWidget {
       _ => colorScheme.primary,
     };
   }
-
-  /// Format relative time for hotspot detection
-  String _formatRelativeTime(DateTime detectedAt) {
-    final now = DateTime.now().toUtc();
-    final difference = now.difference(detectedAt);
-
-    if (difference.inMinutes < 60) {
-      final minutes = difference.inMinutes;
-      return 'Detected $minutes ${minutes == 1 ? 'minute' : 'minutes'} ago';
-    } else if (difference.inHours < 24) {
-      final hours = difference.inHours;
-      return 'Detected $hours ${hours == 1 ? 'hour' : 'hours'} ago';
-    } else if (difference.inDays < 7) {
-      final days = difference.inDays;
-      return 'Detected $days ${days == 1 ? 'day' : 'days'} ago';
-    } else {
-      return 'Detected on ${_formatDateTime(detectedAt)}';
-    }
-  }
 }
 
-/// Section wrapper for clearer grouping of fire details
-class _InfoSection extends StatelessWidget {
-  const _InfoSection({required this.title, required this.children, this.icon});
+/// Key metric row data class for _buildKeyMetricsSection
+class _KeyMetricRow {
+  final IconData icon;
+  final String label;
+  final String value;
+
+  const _KeyMetricRow({
+    required this.icon,
+    required this.label,
+    required this.value,
+  });
+}
+
+/// Expandable "More details" section with smooth animation
+///
+/// Inspired by RiskBanner's ExpandableLocationPanel:
+/// - Chevron icon that rotates on expand/collapse
+/// - Subtle background on header row
+/// - Smooth height animation using AnimatedSize
+/// - Content not built when collapsed (performance)
+class _ExpandableDetailsSection extends StatefulWidget {
+  const _ExpandableDetailsSection({
+    required this.title,
+    required this.children,
+  });
 
   final String title;
   final List<Widget> children;
-  final IconData? icon;
+
+  @override
+  State<_ExpandableDetailsSection> createState() =>
+      _ExpandableDetailsSectionState();
+}
+
+class _ExpandableDetailsSectionState extends State<_ExpandableDetailsSection>
+    with SingleTickerProviderStateMixin {
+  bool _isExpanded = false;
+  late AnimationController _controller;
+  late Animation<double> _iconTurns;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      duration: const Duration(milliseconds: 200),
+      vsync: this,
+    );
+    _iconTurns = Tween<double>(begin: 0.0, end: 0.5).animate(
+      CurvedAnimation(parent: _controller, curve: Curves.easeInOut),
+    );
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  void _handleTap() {
+    setState(() {
+      _isExpanded = !_isExpanded;
+      if (_isExpanded) {
+        _controller.forward();
+      } else {
+        _controller.reverse();
+      }
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
-    final cs = theme.colorScheme;
+    final colorScheme = theme.colorScheme;
     final isDark = theme.brightness == Brightness.dark;
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        Row(
-          children: [
-            if (icon != null) ...[
-              Icon(icon, size: 18, color: cs.onSurfaceVariant),
-              const SizedBox(width: 6),
-            ],
-            Text(
-              title,
-              style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                    fontWeight: FontWeight.w600,
-                    color: cs.onSurfaceVariant,
-                  ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 8),
-        Container(
-          decoration: BoxDecoration(
-            color: isDark ? cs.surfaceContainerHigh : cs.surfaceContainerLowest,
+        // Header row with tap target
+        Semantics(
+          button: true,
+          expanded: _isExpanded,
+          label:
+              '${widget.title}. ${_isExpanded ? "Tap to collapse" : "Tap to expand"}',
+          child: InkWell(
+            onTap: _handleTap,
             borderRadius: BorderRadius.circular(12),
-            border: Border.all(color: cs.outlineVariant),
+            child: Container(
+              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 12),
+              decoration: BoxDecoration(
+                // Subtle background color for header row
+                color: isDark
+                    ? colorScheme.surfaceContainerHigh
+                    : colorScheme.surfaceContainerLow,
+                borderRadius: BorderRadius.only(
+                  topLeft: const Radius.circular(12),
+                  topRight: const Radius.circular(12),
+                  bottomLeft: Radius.circular(_isExpanded ? 0 : 12),
+                  bottomRight: Radius.circular(_isExpanded ? 0 : 12),
+                ),
+                border: Border.all(color: colorScheme.outlineVariant),
+              ),
+              child: Row(
+                children: [
+                  Icon(
+                    Icons.info_outline,
+                    size: 18,
+                    color: colorScheme.onSurfaceVariant,
+                  ),
+                  const SizedBox(width: 8),
+                  Expanded(
+                    child: Text(
+                      widget.title,
+                      style: theme.textTheme.titleSmall?.copyWith(
+                        fontWeight: FontWeight.w600,
+                        color: colorScheme.onSurfaceVariant,
+                      ),
+                    ),
+                  ),
+                  RotationTransition(
+                    turns: _iconTurns,
+                    child: Icon(
+                      Icons.expand_more,
+                      size: 24,
+                      color: colorScheme.onSurfaceVariant,
+                    ),
+                  ),
+                ],
+              ),
+            ),
           ),
-          padding: const EdgeInsets.all(12),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              for (int i = 0; i < children.length; i++) ...[
-                if (i > 0) const SizedBox(height: 12),
-                children[i],
-              ],
-            ],
-          ),
+        ),
+        // Expandable content with smooth animation
+        AnimatedSize(
+          duration: const Duration(milliseconds: 200),
+          curve: Curves.easeInOut,
+          alignment: Alignment.topCenter,
+          child: _isExpanded
+              ? Container(
+                  width: double.infinity,
+                  padding: const EdgeInsets.all(12),
+                  decoration: BoxDecoration(
+                    color: isDark
+                        ? colorScheme.surfaceContainerHigh
+                        : colorScheme.surfaceContainerLowest,
+                    borderRadius: const BorderRadius.only(
+                      bottomLeft: Radius.circular(12),
+                      bottomRight: Radius.circular(12),
+                    ),
+                    border: Border(
+                      left: BorderSide(color: colorScheme.outlineVariant),
+                      right: BorderSide(color: colorScheme.outlineVariant),
+                      bottom: BorderSide(color: colorScheme.outlineVariant),
+                    ),
+                  ),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      for (int i = 0; i < widget.children.length; i++) ...[
+                        if (i > 0) const SizedBox(height: 4),
+                        widget.children[i],
+                      ],
+                    ],
+                  ),
+                )
+              : const SizedBox.shrink(),
         ),
       ],
     );
